requirements.txt
Fabric==3.2.2
fastapi==0.109.0
generic_connection_pool==0.7.0
passlib==1.7.4
pydantic==2.5.3
python-dotenv==1.0.0
python_jose==3.3.0
typing_extensions==4.9.0
uvicorn==0.26.0
python-multipart==0.0.6
jaydebeapi==1.2.3
PyYAML==6.0.2
gitdb==4.0.12
GitPython==3.1.44
